+++ src/types/responses/sabnzbd/QueueResponse.ts
@@ -1,4 +1,4 @@
-import { formatBytes } from '../../../utils/Utils';
+import { formatBytes } from '../../../utils/formatters';
 
 export enum QueueStatus {
     DOWNLOADING = 'Downloading',
