+++ src/utils/formatters.ts
@@ -1,0 +1,9 @@
+export function formatBytes(bytes: number, unit: boolean = true, decimals: number = 2): string {
+    if (bytes === 0) return '0 Bytes';
+
+    const k: number = 1024;
+    const sizes: string[] = ['Bytes', 'KB', 'MB', 'G', 'TB', 'PB'];
+    const i: number = Math.floor(Math.log(bytes) / Math.log(k));
+
+    return parseFloat((bytes / Math.pow(k, i)).toFixed(decimals)) + (unit ? ' ' + sizes[i] : '');
+}
