+++ tests/service/schedule/NativeScheduleService.test.ts
@@ -1,7 +1,7 @@
 // __tests__/NativeScheduleService.test.ts
 import axios from 'axios';
 
-import * as Utils from '../../../src//utils/Utils';
+import * as Utils from '../../../src/utils/Utils';
 import configService from '../../../src/service/configService';
 import iplayerDetailsService from '../../../src/service/iplayerDetailsService';
 import loggingService from '../../../src/service/loggingService';
@@ -14,11 +14,7 @@ jest.mock('../../../src/service/iplayerDetailsService');
 jest.mock('../../../src/service/loggingService');
 jest.mock('../../../src/service/redis/redisCacheService');
 jest.mock('../../../src/service/search/NativeSearchService');
-jest.mock('../../../src/utils/Utils', () => ({
-    ...jest.requireActual('../../../src/utils/Utils'),
-    getQualityProfile: jest.fn(),
-    splitArrayIntoChunks: jest.fn(),
-}));
+jest.mock('../../../src/utils/Utils');
 
 describe('NativeScheduleService', () => {
     beforeEach(() => {
